+++ tests/test_handlers.py
@@ -119,7 +119,7 @@ async def _run_pr_handler(result: ReviewResult | Exception, mock_engine_cls, moc
 
     with (
         patch("mira.dashboard.api._app_db") as mock_db,
-        patch("mira.notifications.dispatch_event", new_callable=AsyncMock) as mock_dispatch,
+        patch("mira.outbound_webhooks.dispatch_event", new_callable=AsyncMock) as mock_dispatch,
     ):
         mock_db.get_repo.return_value = MagicMock(status="ready")  # indexed → skip note
         await handle_pull_request(_make_pr_payload(), mock_app_auth, "mira-bot")
